backend/metadata_writer/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=- -
#
absl-py==0.12.0           # via ml-metadata
attrs==20.3.0             # via ml-metadata
cachetools==5.0.0         # via google-auth
certifi==2021.10.8        # via kubernetes, requests
charset-normalizer==2.0.10  # via requests
google-auth==2.4.1        # via kubernetes
grpcio==1.43.0            # via ml-metadata
idna==3.3                 # via requests
kubernetes==10.1.0        # via -r -
lru-dict==1.1.7           # via -r -
ml-metadata==1.5.0        # via -r -
oauthlib==3.1.1           # via requests-oauthlib
protobuf==3.19.3          # via ml-metadata
pyasn1-modules==0.2.8     # via google-auth
pyasn1==0.4.8             # via pyasn1-modules, rsa
python-dateutil==2.8.2    # via kubernetes
pyyaml==3.13              # via kubernetes
requests-oauthlib==1.3.0  # via kubernetes
requests==2.27.1          # via kubernetes, requests-oauthlib
rsa==4.8                  # via google-auth
six==1.16.0               # via absl-py, google-auth, grpcio, kubernetes, ml-metadata, python-dateutil
urllib3==1.26.8           # via kubernetes, requests
websocket-client==1.2.3   # via kubernetes
tabulate
# The following packages are considered to be unsafe in a requirements file:
# setuptools
